refactor(数据导入): 实现数据源配置和列映射的属性变更通知
将IDataSourceConfig接口和ColumnMapping类实现INotifyPropertyChanged接口
添加属性变更通知支持，优化UI控件的双向绑定
统一修改数据源类型比较为整型值比较

+++ RUINORERP.UI/SysConfig/BasicDataImport/DynamicDataValidator.cs
@@ -108,8 +108,8 @@ namespace RUINORERP.UI.SysConfig.BasicDataImport
                 return false;
 
             // 使用 DataSourceType 判断，不再依赖字符串匹配
-            return mapping.ColumnDataSourceType == DataSourceType.SystemGenerated
-                || mapping.ColumnDataSourceType == DataSourceType.DefaultValue;
+            return mapping.ColumnDataSourceType == (int)DataSourceType.SystemGenerated
+                || mapping.ColumnDataSourceType == (int)DataSourceType.DefaultValue;
         }
 
         /// <summary>
@@ -132,7 +132,7 @@ namespace RUINORERP.UI.SysConfig.BasicDataImport
                     PropertyInfo property = entityType.GetProperty(mapping.SystemField?.Key);
                     if (property == null) continue;
 
-                    if (mapping.ColumnDataSourceType == DataSourceType.ForeignKey)
+                    if (mapping.ColumnDataSourceType == (int)DataSourceType.ForeignKey)
                     {
                         // 外键验证
                         string foreignKeyError;
@@ -464,7 +464,7 @@ namespace RUINORERP.UI.SysConfig.BasicDataImport
         private void ValidateForeignKeys(DataTable dataTable, List<ColumnMapping> mappings, List<ValidationError> errors)
         {
             // 直接使用 DataSourceType 筛选外键映射
-            var foreignKeyMappings = mappings.Where(m => m.ColumnDataSourceType == DataSourceType.ForeignKey).ToList();
+            var foreignKeyMappings = mappings.Where(m => m.ColumnDataSourceType == (int)DataSourceType.ForeignKey).ToList();
             if (!foreignKeyMappings.Any())
             {
                 return;
